refactor: split BacktestNeuralNet out of dynamic_plot_backtest (roadmap 5.3)
Move the orchestrator BacktestNeuralNet (composes the loss/accuracy/perf
dynamic plots + drives the backtest figure) into a new
backtest/backtest_neural_net.py, separating it from the plot primitives.
The tightly-coupled DynaPlot* inheritance family (DynaPlotBackTest base +
Accuracy/Loss/Perf) stays together. Re-exported from fynance.backtest;
existing imports (incl. models.rolling) preserved. Sphinx xrefs updated.
