+++ fynance/backtest/__init__.py
@@ -16,7 +16,14 @@ Some tools to backtest strategies
 
 """
 
-from . import dynamic_plot_backtest, plot_backtest, plot_tools, print_stats
+from . import (
+    backtest_neural_net,
+    dynamic_plot_backtest,
+    plot_backtest,
+    plot_tools,
+    print_stats,
+)
+from .backtest_neural_net import *
 from .dynamic_plot_backtest import *
 from .plot_backtest import *
 from .plot_tools import *
@@ -26,3 +33,4 @@ __all__ = print_stats.__all__
 __all__ += plot_tools.__all__
 __all__ += plot_backtest.__all__
 __all__ += dynamic_plot_backtest.__all__
+__all__ += backtest_neural_net.__all__
